fix(core): fix native class constructor method
This commit fixes the native class constructor, namely the constructor doesn't
have the `this` value to hand until the native object is built. This was causes derefering panics since the *target*
was being passed as `this` -- which doesn't have the same type as the instance type.
The solution is to split the constructor into a *data* constructor
and an *object* constructor.