fix(proto): remove `source` from `Operation`
`source` is not only a redundant field in
`Operation` (since `public_key` should be
able to derive `source`) but there are several
bugs in the verification logic (around source vs
public_key) which result in criticial security vulns.
This commit removes `source` entirely, `public_key` is the source of truth for the
signer's identity.