feat(proto): parsed code for v2 runtime
Checks in ParsedCode should contain all the checks that could reject a code before it is deployed as a smart function.
Note that even if code is parsable, it might not necesarily mean that it is a valid smart function. It is also possible
for malicious to be run. Specifically, the user could try to modify state through KV or fetch. For this reason, the
JstzRuntime instantiated is built w no access to the protocol or related ops - fetch, kv, console. The code could also
run an infinite loop which will need gas to be prevented.

Parsing involves compiling the source, instantiating module imports, evaluating the module code and checking that the
default handler exists. Note that we do no not allow module imports so existence of imports will return error